report_parser: Added optional dependencies
This patch updates the `dependencies_mdt_collect` method
to support unexpected dependency-containing directories.

The added behaviour enables third party tools and libraries
to be added to the `tf-m-dependencies-checkout` for use in
board specific pipelines.

The new behaviour, will search for the .git directory, or a
key-value version.info file, in the user defined path, and
launch the appropriate parsing logic. On exception it will
still add a {"info": "N.A"} tag and not fail the job.

Change-Id: I356f6b86efa0f7e0818c093415afee8d5a492e60
Signed-off-by: Minos Galanakis <minos.galanakis@linaro.org>