PSoC64: Add build combinations into "full" group
The old internal CI will run "full" group for combinations
parse, this will add 4 from PSoC64 to the total number:
	psoc64_ARMCLANG_ConfigRegressionIPC_Release_NOBL2
	psoc64_GNUARM_ConfigRegressionIPC_Release_NOBL2
	psoc64_ARMCLANG_ConfigRegressionIPCTfmLevel2_Release_NOBL2
	psoc64_GNUARM_ConfigRegressionIPCTfmLevel2_Release_NOBL2

As the TFM CI on tf.org does not support Clang at the moment, align the
group "full_gnuarm" simultaneously for combinations:
	psoc64_GNUARM_ConfigRegressionIPC_Release_NOBL2
	psoc64_GNUARM_ConfigRegressionIPCTfmLevel2_Release_NOBL2

These groups may not be used for the future when the new CI fully runs on
tf.org. But worth keeping them for the time being.

Change-Id: I689e82403224857d854b4d46cd076fef27dce54d
Signed-off-by: Karl Zhang <karl.zhang@arm.com>