create a coverity builder
The Coverity Scanner requires a 'platform builder' which in this case,
it is a makefile (cov-makefile) that take cares of platform
compilation under the trusted-firmware-m folder. The compiled files
are the only source files scanned by Coverity, thus the more platforms
included on the platform builder, the better is the static check
scan. So far, the only platform considered is mps2/an521 but much more
are desired as explained above.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: If8ce6a70e8568c5e8b0116555b20377855ac9a55