CopyDocs: limit archiving doc artifacts
The artifacts from doc job were needed as a part of nightly artifacts
because https://www.trustedfirmware.org/ have pointed to nightly job for
documentation.

The CI has a doc nightly job for doc only "tf-m-build-docs-nightly"
which has updated as new link for documentation on tf.org.
As the copy docs consume a lot of time and disk space without any
purpose, so, remove the copy for nightly job.

But, the per patch job(tf-m-build-and-test) need the doc copy, because
of unmerged changes may affect the documentation build.

Signed-off-by: Karl Zhang <karl.zhang@arm.com>
Change-Id: I2c30ee237d07627bb5b656cc6dfa142003c7006f