CI Artifacts: Archive Build Artifacts in Original Path
TF-M has transferred to new build system.
Build artifacts paths are also different from the old one.

To keep the same URL to build artifacts,
which is needed by LAVA server,
artifact paths need to be same to the old one.
So build artifacts should be archived in original path.

Change-Id: I6729370bf1b5f2160a8dd41934f55ca4970ae7c3
Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>