run-build.sh: workaround armclang error: Failed to check out a license.
armclang --version
armclang: error: Failed to check out a license.
The license file could not be found. Check that ARMLMD_LICENSE_FILE is set
correctly.

ARMLMD_LICENSE_FILE should be set in the TF-M docker image:
https://git.trustedfirmware.org/ci/dockerfiles.git/commit/?id=07e1447b0f1950045112f8f9e88f6fd68b1d7653

Until the image is deployed, use a workaround to ensure the environment
variable is set. The workaround is limited to trustedfirmware.org domain
to avoid side effect.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>
Change-Id: I6b732fe5b5a0a8c5a9abae1e2b5c1280358486d2