lava_create_jobs: generate_test_definitions: relax access to config dict
Relax the access to config dictionary by using get() method instead of
square brackets, in order to avoid to raise KeyError exceptions.
It's required for Musca support because the current code base is tightly
coupled with the existing supported platform like FVP and MPS2.
Some params/config doesn't exist for Musca, they apply only to FVP/MPS2,
and we want to gracefully continue if they're missing.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>
Change-Id: I0cd16abcc3e6b41a9fcebfae418c7bfbf162bf36