Switch all platforms artifacts to use the new build system
TF-M changed binaries name and location since 1.1 release.
After enabling additional platforms (Musca B1 and MPS2 on QEMU),
the infra health check started to fail. The build configs were broken in
the first place, due to workarounds to support mixing old and new build
system. This patch finished the migration to the TF-M new build system
and get rid of the related workarounds, FIXMEs and TODOs.
It requires a new recovery image with an updated images.txt file to
match the new filenames.

* don't run cmake install commands. It isn't necessary to install the
  binaries under the deprecated paths. Binaries are installed under bin,
  out of the box.
* Adjust archiveArtifacts to build/bin/**. Files under build/install/**
  aren't necessary anymore.
* Adjust get_artifact_url() to return bin/ URL.
* Add comments to lava_helper_configs.py to clarify the binaries to use.
* Update the binaries filenames. mcuboot.* and tfm_sign.bin are gone.
* Fix qemu_mps2: nobl2 -> bl2. Rename accordingly.
* Cleanup run-build.sh as we don't need to remove/rename/reshuffle files
  anymore.
* Adjust fastmodel_wrapper to the binaries paths and filenames used by
  the new build system.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>
Change-Id: I8c23cf1d5430b21bb7ee5b232a9125b089792efd