LAVA: Disable tests on QEMU
Tests for same build configs are likely to pass on fvp & pyhsical boards
but fail on QEMU, which would get TF-M Open CI blocked.
This is a bug in QEMU. It is likely to be fixed in the coming release.
Disable tests on QEMU for the time being to avoid blocking CI.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I2552611d69ba8d601bd98b374c842e0729853ac1