ci.jpl: Pass CI_SCRIPTS_REPO param to downstream jobs
Just as CI_SCRIPTS_BRANCH is already passed. This is required for seamless
testing on staging, where only parent job needs CI_SCRIPTS_REPO value to
be patched, and then just passes it to the downstream jobs. Before that,
the builds were broken, because personal CI_SCRIPTS_BRANCH value was passed,
but CI_SCRIPTS_REPO wasn't.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: Ice64f0b94ef7fb3be568449d1589a94b952e7c70