lava_wait_jobs.py: Retry only fetching LAVA jobs
Previously, the whole functions wait_for_jobs() function was wrapped in
the retry logic, but it contained not just LAVA logic, but also reporting
logic. If there was an exception in reporting logic, the whole process
of getting LAVA results was repeated, only leading to confusing output
in the logs and extra LAVA load in case of code errors. Split the
wait_for_jobs() in two functions and retry only the actual LAVA logic.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: Iae73d3dc6947ffe1fd12791f01989b4ce44d65e1