LAVA Helper: Remove 'required' item in LAVA monitors
In the current logic of LAVA monitors, 'required' items are not used.
According to the jinja2 templates, 'required' is not contained in LAVA
configs when generating LAVA job definations.

The monitor starts and ends correspondingly once the 'start' and 'end'
strings are matched in the logs.
The test result is parsed from the 'fixup' pattern printed in the log.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I15ff734d8daa1942f88ee3a27acd4633211a6fe5