lava-submit.jpl: Rework detection of "per-patch" runs
As those are defined as started by the tf-m-static parent job, check just
for that. Checking for tf-m-build-and-test doesn't make sense, as that job
is run for manual testing purposes, when adhoc filtering is not required
and confusing.

When checking for the job name, using .endsWith() method, to accommodate
per-user job copies on staging (user-tf-m-static).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I82475fbcc8719abefa92b389283e240be2c9dd30