LAVA: Update start string in reg test monitor
Use "Execute test suites" as the start string in reg test monitor,
which will be surely printed in all configs with reg tests.

"Starting" would only be printed in Debug mode. So it will lead to
unexpected time out error in Release & Minsize mode, because the
monitor cannot match the start string.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I2adf06387676c486670442b786f3947bec8efcf5