Build: Support multi params in 'extra_params'
This patch extend 'extra_params' in build helper to support multiple
extra params in one config. Params are seperated by ', '.

For example, for a config with both FPHARD and LZOFF enabled,
extra_params can be set as "FPHARD, LZOFF". Then the build helper
will append the configs in the build command generated automatically.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I645284c4ef37d47feb5e31fe093d67a446778c5a