Rewrite logics to generate job results summary
This patch implements a report csv helper, which can be shared between
build and test jobs to generate csv files for result summary. It makes
the scripts easier to maintain.

This patch also aligns the data structures of build and test results, to
simplify the logic for email notification and other artifacts.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I61d3f45933908e31880bde0de3250d3066ca79e8