fvp_mps2.jinja2: Increase CPU limit to 600
This is required for the TFM_NS_IRQ_TEST_FLIH_1101 testcase, which takes
long time when run with code coverage plugin. Such long, that previously
it seemed that it simply hanged. My thorough testing shows that it
eventually seems to finish, just takes 8-9 minutes. So, set CPU limit
to 10 minutes and also increase simulation limit slightly above that.

This is related to https://linaro.atlassian.net/browse/TFC-364

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I1e2799fa581b26c3bc5ef0b78325b1a1c664cef2