ci.jpl: Remove not useful and confusing set +x/-x commands
This partially reverts commit d091f811a12. That commit intended to avoid to
"print duplicated Python commands in CI logs", but the actual result of it
is that logs now contain stuff like:

[Pipeline] sh
00:25:11  + set +x
[Pipeline] sh
00:25:12  + set +x
[Pipeline] sh
00:25:12  + set +x
[Pipeline] sh
00:25:13  + set +x
[Pipeline] sh
00:25:13  + set +x
[Pipeline] sh
00:25:14  + set +x
[Pipeline] sh
00:25:14  + set +x
[Pipeline] sh
00:25:15  + set +x
(may be repeated 100 of times).

It's much better to print the actual Python command. Actually, it's
important to print the actual commands being run, that's immense help
to debug any issues with CI jobs (which are many, and even when they
won't be many, they will still happen).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I5c66ae7cc831f0a5d403700d3b1260061e09fedb