lava_wait_jobs.py: Initialize and use logging
How this works now, that there're visible build progress during compilation,
etc., but then jobs hang literally for hours in a call to lava_wait_jobs.py.

That's not ok, there should be regular processing updates visible to a human
operator, so they understand what happens, be able to debug issues, and look
for ways to improve. As the output from the script contains job results and
being parsed like that, an easy way to output diagnostics would be to stderr.
This is automatically achieved with Python logging module, so just use that.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I44d11fbc95848d9cdd652e1cb0602e3ed0da09ec